+++ lib/odd_job/supervisor.ex
@@ -1,6 +1,28 @@
 defmodule OddJob.Supervisor do
-  @moduledoc false
+  @moduledoc """
+  The `OddJob.Supervisor` is responsible for starting and supervising a job pool.
+
+  All of this module's public functions can be called using the `OddJob` namespace. See
+  the `OddJob` documentation for usage.
+  """
+  @moduledoc since: "0.1.0"
   use Supervisor
+
+  @type start_arg :: atom | {atom, [start_option]} | [start_option]
+  @type start_option ::
+          {:name, atom}
+          | {:pool_size, non_neg_integer}
+          | {:max_restarts, non_neg_integer}
+          | {:max_seconds, non_neg_integer}
+  @type child_spec :: %{
+          id: atom,
+          start: {OddJob.Supervisor, :start_link, [start_arg]},
+          type: :supervisor
+        }
+
+  @doc false
+  @spec start_link(start_arg) :: Supervisor.on_start()
+  def start_link(start_arg)
 
   def start_link(name) when is_atom(name) do
     start_link(name: name)
@@ -23,6 +45,7 @@ defmodule OddJob.Supervisor do
     Supervisor.init(children, strategy: :one_for_one)
   end
 
+  @doc false
   def child_spec(name) when is_atom(name), do: child_spec(name: name)
 
   def child_spec({name, opts}) when is_atom(name) and is_list(opts) do
@@ -37,6 +60,7 @@ defmodule OddJob.Supervisor do
     |> Supervisor.child_spec(id: id(opts[:name]))
   end
 
+  @doc false
   @spec pool_id(atom | binary) :: atom
   def pool_id(name) when is_atom(name) or is_binary(name), do: :"#{name}_pool"